+++ packages/flutter_vless_platform_interface/lib/flutter_vless_platform_interface.dart
@@ -2,4 +2,5 @@ library;
 
 export 'src/vless_platform.dart';
 export 'src/vless_status.dart';
+export 'src/vless_method_channel_adapter.dart';
 export 'src/method_channel_vless_platform.dart';